- feature_#098_DiaryRepository 인터페이스 생성
